requirements.txt
# Core dependency (optional - falls back to urllib if not installed)
anthropic>=0.39.0

# Phase 1 & Phase 2 converter dependencies
pyyaml>=6.0
jinja2>=3.0

# Development dependencies
pytest>=7.0
pytest-cov>=4.0
